fix: transfer input in shared and fix lint errors
