+++ src/components/Shared/Input/Input.types.ts
@@ -1,10 +1,9 @@
-import { ChangeEvent, ReactNode } from 'react';
 import { ChangeEventHandler } from 'react';
 
 export interface InputProps {
-  id: string,
+  id: string;
   placeholder: string;
   inputClassName?: string;
-  primary?: boolean,
+  primary?: boolean;
   onChange?: ChangeEventHandler<HTMLInputElement> | undefined;
 }
